added comments functions to the app.py api file, should work

+++ photopro-app/api/utils/database/comments.py
@@ -45,10 +45,10 @@ def post_comment_to_comment(image_id, commenter, comment, reply_id, conn, cur):
         return False
 
 
-def delete_comment(comment_id, conn, cur):
+def delete_comment(comment_id, user_id, conn, cur):
     try:
         cur.execute('SAVEPOINT save_point')
-        cmd = "DELETE FROM comments WHERE comment_id={}".format(comment_id)
+        cmd = "DELETE FROM comments WHERE comment_id={} AND commenter={}".format(comment_id, user_id)
         cur.execute(cmd)
         conn.commit()
         return True
@@ -69,10 +69,20 @@ def delete_comment(comment_id, conn, cur):
 def get_comments_to_image(image_id, batch_size, conn, cur):
     try:
         cur.execute('SAVEPOINT save_point')
-        cmd = "SELECT * FROM comments WHERE image_id={} AND reply_id is null LIMIT {}".format(image_id, batch_size)
+        cmd = "SELECT comment_id, image_id, commenter, comment, reply_id, created_at FROM comments WHERE image_id={} " \
+              "AND reply_id is null LIMIT {}".format(image_id, batch_size)
+        print(cmd)
         cur.execute(cmd)
         conn.commit()
-        return True
+        data = cur.fetchmany(batch_size)
+
+        length = len(data)
+        # print(length)
+        if length == 0:
+            return False
+        else:
+            # print(data)
+            return data
     except psycopg2.errors.UniqueViolation as e:
         print(e)
         cur.execute('ROLLBACK TO SAVEPOINT save_point')
@@ -90,10 +100,20 @@ def get_comments_to_image(image_id, batch_size, conn, cur):
 def get_comments_to_comment(reply_id, batch_size, conn, cur):
     try:
         cur.execute('SAVEPOINT save_point')
-        cmd = "DELETE FROM comments WHERE reply_id={} LIMIT {}".format(reply_id, batch_size)
+        cmd = "SELECT comment_id, image_id, commenter, comment, reply_id, created_at FROM comments WHERE reply_id={} " \
+              "LIMIT {}".format(reply_id, batch_size)
+        print(cmd)
         cur.execute(cmd)
         conn.commit()
-        return True
+        data = cur.fetchmany(batch_size)
+
+        length = len(data)
+        # print(length)
+        if length == 0:
+            return False
+        else:
+            # print(data)
+            return data
     except psycopg2.errors.UniqueViolation as e:
         print(e)
         cur.execute('ROLLBACK TO SAVEPOINT save_point')
